backend/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
email-validator==2.0.0
itsdangerous==2.1.2
bcrypt==4.0.1
PyJWT==2.8.0